Removed debug logging
